test: all project tests have been adjusted
Especially the tests that are part of the project's authentication module.
