+++ api_inmobiliaria/tests/authentication/views/jwt/test_update.py
@@ -59,25 +59,23 @@ class TestUpdateTokensAPIView:
         base_user, _, _ = self.user_factory.user(
             user_role=user_role, active=True, save=True, add_perm=False
         )
-        jwt_data = self.jwt_factory.access_and_refresh(
+        access_token_data = self.jwt_factory.access(
             user_role=base_user.content_type.model,
             user=base_user,
-            exp_access=True,
-            exp_refresh=False,
+            exp=True,
             save=True,
         )
 
         # Simulating the request
         response = self.client.post(
             path=self.path,
-            data=jwt_data["tokens"],
+            data={"access_token": access_token_data["token"]},
             content_type="application/json",
         )
 
         # Asserting that response data is correct
         assert response.status_code == status.HTTP_200_OK
         assert "access_token" in response.data
-        assert "refresh_token" in response.data
 
     def test_if_empty_data(self) -> None:
         """
@@ -85,8 +83,7 @@ class TestUpdateTokensAPIView:
         when the request data is empty.
         """
 
-        error_messages_expected = {
-            "refresh_token": [DEFAULT_ERROR_MESSAGES["required"]],
+        error_message_expected = {
             "access_token": [DEFAULT_ERROR_MESSAGES["required"]],
         }
 
@@ -107,19 +104,14 @@ class TestUpdateTokensAPIView:
             for field, errors in response.data["detail"].items()
         }
 
-        for field, message in error_messages_expected.items():
+        for field, message in error_message_expected.items():
             assert response_errors_formated[field] == message
 
     @pytest.mark.parametrize(
         argnames="data, error_message",
         argvalues=[
             (
-                {
-                    "access_token": JWTFactory.access_invalid(),
-                    "refresh_token": JWTFactory.refresh(
-                        exp=False, save=False
-                    ).get("token"),
-                },
+                {"access_token": JWTFactory.access_invalid()},
                 JWTErrorMessages.INVALID_OR_EXPIRED.value.format(
                     token_type="access"
                 ),
@@ -127,56 +119,15 @@ class TestUpdateTokensAPIView:
             (
                 {
                     "access_token": JWTFactory.access(
-                        exp=True, save=False  # fmt: off
-                    ).get("token"),
-                    "refresh_token": JWTFactory.refresh_invalid(),
-                },
-                JWTErrorMessages.INVALID_OR_EXPIRED.value.format(
-                    token_type="refresh"
-                ),
-            ),
-            (
-                {
-                    "access_token": JWTFactory.access(
-                        exp=False, save=False
-                    ).get("token"),
-                    "refresh_token": JWTFactory.refresh(
                         exp=False, save=False
                     ).get("token"),
                 },
                 JWTErrorMessages.ACCESS_NOT_EXPIRED.value,
             ),
-            (
-                {
-                    "access_token": JWTFactory.access(
-                        exp=True, save=False  # fmt: off
-                    ).get("token"),
-                    "refresh_token": JWTFactory.refresh(
-                        exp=True, save=False
-                    ).get("token"),
-                },
-                JWTErrorMessages.INVALID_OR_EXPIRED.value.format(
-                    token_type="refresh"
-                ),
-            ),
-            (
-                {
-                    "access_token": JWTFactory.access(
-                        exp=True, save=False  # fmt: off
-                    ).get("token"),
-                    "refresh_token": JWTFactory.refresh(
-                        exp=False, save=False
-                    ).get("token"),
-                },
-                JWTErrorMessages.USER_NOT_MATCH.value,
-            ),
         ],
         ids=[
             "access_token_invalid",
-            "refresh_token_invalid",
             "access_token_not_expired",
-            "refresh_token_expired",
-            "user_not_match",
         ],
     )
     def test_if_token_validation_failed(
@@ -204,31 +155,7 @@ class TestUpdateTokensAPIView:
         assert response.data["code"] == code_expected
         assert response.data["detail"] == error_message
 
-    @pytest.mark.parametrize(
-        argnames="access_blacklist, refresh_blacklist, error_message",
-        argvalues=[
-            (
-                True,
-                False,
-                JWTErrorMessages.BLACKLISTED.value.format(token_type="access"),
-            ),
-            (
-                False,
-                True,
-                JWTErrorMessages.BLACKLISTED.value.format(token_type="refresh"),
-            ),
-        ],
-        ids=[
-            "access_token_blacklisted",
-            "refresh_token_blacklisted",
-        ],
-    )
-    def test_if_token_blacklisted(
-        self,
-        access_blacklist: bool,
-        refresh_blacklist: bool,
-        error_message: str,
-    ) -> None:
+    def test_if_token_blacklisted(self) -> None:
         """
         This test is responsible for validating the expected behavior of the view
         when the JWTs are blacklisted.
@@ -238,102 +165,33 @@ class TestUpdateTokensAPIView:
         base_user, _, _ = self.user_factory.searcher_user(
             active=True, save=True, add_perm=False
         )
-        access_token = self.jwt_factory.access(
+        access_token_data = self.jwt_factory.access(
             user_role=base_user.content_type.model,
             user=base_user,
+            add_blacklist=True,
             exp=True,
             save=True,
-            add_blacklist=access_blacklist,
-        ).get("token")
-        refresh_token = self.jwt_factory.refresh(
-            user_role=base_user.content_type.model,
-            user=base_user,
-            exp=False,
-            save=True,
-            add_blacklist=refresh_blacklist,
-        ).get("token")
+        )
 
         # Simulating the request
         response = self.client.post(
             path=self.path,
-            data={"access_token": access_token, "refresh_token": refresh_token},
+            data={"access_token": access_token_data["token"]},
             content_type="application/json",
         )
 
         # Asserting that response data is correct
         status_code_expected = JWTAPIError.status_code
         code_expected = JWTAPIError.default_code
+        error_message_expected = JWTErrorMessages.BLACKLISTED.value.format(
+            token_type="access"
+        )
 
         assert response.status_code == status_code_expected
         assert response.data["code"] == code_expected
-        assert response.data["detail"] == error_message
+        assert response.data["detail"] == error_message_expected
 
-    def test_if_user_not_match(self) -> None:
-        """
-        This test is responsible for validating the expected behavior of the view
-        when the user does not match the JWT data.
-        """
-
-        # Creating the JWTs to be used in the test
-        access_token = self.jwt_factory.access(
-            exp=True, save=False  # fmt: off
-        ).get("token")
-        refresh_token = self.jwt_factory.refresh(
-            exp=False, save=False  # fmt: off
-        ).get("token")
-
-        # Simulating the request
-        response = self.client.post(
-            path=self.path,
-            data={"access_token": access_token, "refresh_token": refresh_token},
-            content_type="application/json",
-        )
-
-        # Asserting that response data is correct
-        status_code_expected = JWTAPIError.status_code
-        code_expected = JWTAPIError.default_code
-        response_data_expected = JWTErrorMessages.USER_NOT_MATCH.value
-
-        assert response.status_code == status_code_expected
-        assert response.data["code"] == code_expected
-        assert response.data["detail"] == response_data_expected
-
-    def test_if_token_not_found(self) -> None:
-        """
-        This test is responsible for validating the expected behavior of the view
-        when the JWTs are not found in the database.
-        """
-
-        # Creating the JWTs to be used in the test
-        base_user, _, _ = self.user_factory.searcher_user(
-            active=True, save=True, add_perm=False
-        )
-        jwt_data = self.jwt_factory.access_and_refresh(
-            user_role=base_user.content_type.model,
-            user=base_user,
-            exp_access=True,
-            exp_refresh=False,
-            save=False,
-        )
-
-        # Simulating the request
-        response = self.client.post(
-            path=self.path,
-            data=jwt_data["tokens"],
-            content_type="application/json",
-        )
-
-        # Asserting that response data is correct
-        status_code_expected = ResourceNotFoundAPIError.status_code
-        message = JWTErrorMessages.TOKEN_NOT_FOUND.value
-        response_code_expected = message["code"]
-        response_data_expected = message["detail"].format(token_type="refresh")
-
-        assert response.status_code == status_code_expected
-        assert response.data["code"] == response_code_expected
-        assert response.data["detail"] == response_data_expected
-
-    @patch("apps.users.infrastructure.views.jwt.UserRepository")
+    @patch(target="apps.authentication.infrastructure.views.jwt.UserRepository")
     def test_if_user_not_found(self, user_repository_mock: Mock) -> None:
         """
         This test is responsible for validating the expected behavior of the view
@@ -345,18 +203,17 @@ class TestUpdateTokensAPIView:
         get_base_data.return_value = empty_queryset(model=BaseUser)
 
         # Creating the JWTs to be used in the test
-        jwt_data = self.jwt_factory.access_and_refresh(
-            user=BaseUser(),
+        access_token_data = self.jwt_factory.access(
             user_role="AnyUser",
-            exp_access=True,
-            exp_refresh=False,
+            user=BaseUser(),
+            exp=True,
             save=False,
         )
 
         # Simulating the request
         response = self.client.post(
             path=self.path,
-            data=jwt_data["tokens"],
+            data={"access_token": access_token_data["token"]},
             content_type="application/json",
         )
 
@@ -370,7 +227,7 @@ class TestUpdateTokensAPIView:
         assert response.data["code"] == response_code_expected
         assert response.data["detail"] == response_data_expected
 
-    @patch("apps.users.infrastructure.views.jwt.UserRepository")
+    @patch(target="apps.authentication.infrastructure.views.jwt.UserRepository")
     def test_if_conection_db_failed(self, user_repository_mock: Mock) -> None:
         """
         Test that validates the expected behavior of the view when the connection to
@@ -382,18 +239,17 @@ class TestUpdateTokensAPIView:
         get_base_data.side_effect = DatabaseConnectionAPIError
 
         # Creating the JWTs to be used in the test
-        jwt_data = self.jwt_factory.access_and_refresh(
-            user=BaseUser(),
+        access_token_data = self.jwt_factory.access(
             user_role="AnyUser",
-            exp_access=True,
-            exp_refresh=False,
+            user=BaseUser(),
+            exp=True,
             save=False,
         )
 
         # Simulating the request
         response = self.client.post(
             path=self.path,
-            data=jwt_data["tokens"],
+            data={"access_token": access_token_data["token"]},
             content_type="application/json",
         )
